huazie_Flea请求和过滤器任务代码优化
