+++ flea-frame-core/src/main/java/com/huazie/frame/core/request/FleaRequestXmlDigesterHelper.java
@@ -1,18 +1,18 @@
-package com.huazie.frame.core.filter;
+package com.huazie.frame.core.request;
 
 import com.huazie.frame.common.XmlDigesterHelper;
 import com.huazie.frame.common.util.StringUtils;
 import com.huazie.frame.core.common.FleaCoreConstants;
-import com.huazie.frame.core.filter.config.FilterTask;
-import com.huazie.frame.core.filter.config.FilterTaskChain;
-import com.huazie.frame.core.filter.config.FilterTasks;
-import com.huazie.frame.core.filter.config.FleaRequest;
-import com.huazie.frame.core.filter.config.FleaRequestFilter;
-import com.huazie.frame.core.filter.config.FleaSession;
-import com.huazie.frame.core.filter.config.FleaUrl;
-import com.huazie.frame.core.filter.config.Property;
-import com.huazie.frame.core.filter.config.RedirectUrl;
-import com.huazie.frame.core.filter.config.UrlPrefix;
+import com.huazie.frame.core.request.config.FilterTask;
+import com.huazie.frame.core.request.config.FilterTaskChain;
+import com.huazie.frame.core.request.config.FilterTasks;
+import com.huazie.frame.core.request.config.FleaRequest;
+import com.huazie.frame.core.request.config.FleaRequestFilter;
+import com.huazie.frame.core.request.config.FleaSession;
+import com.huazie.frame.core.request.config.FleaUrl;
+import com.huazie.frame.core.request.config.Property;
+import com.huazie.frame.core.request.config.RedirectUrl;
+import com.huazie.frame.core.request.config.UrlPrefix;
 import org.apache.commons.digester.Digester;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
