+++ flea-frame-core/src/main/java/com/huazie/frame/core/request/config/FilterTask.java
@@ -1,0 +1,48 @@
+package com.huazie.frame.core.request.config;
+
+import org.apache.commons.lang.builder.ToStringBuilder;
+
+/**
+ * <p> flea-request-filter.xml 配置 {@code <filter-task> </filter-task>} 节点 </p>
+ *
+ * @author huazie
+ * @version 1.0.0
+ * @since 1.0.0
+ */
+public class FilterTask {
+
+    private String clazz; // 过滤器任务实现类
+
+    private int order; // 过滤器任务执行顺序
+
+    private String desc; // 过滤器任务描述
+
+    public String getClazz() {
+        return clazz;
+    }
+
+    public void setClazz(String clazz) {
+        this.clazz = clazz;
+    }
+
+    public int getOrder() {
+        return order;
+    }
+
+    public void setOrder(int order) {
+        this.order = order;
+    }
+
+    public String getDesc() {
+        return desc;
+    }
+
+    public void setDesc(String desc) {
+        this.desc = desc;
+    }
+
+    @Override
+    public String toString() {
+        return ToStringBuilder.reflectionToString(this);
+    }
+}
